+++ app/views/admin/borrowings/_borrowing.html.erb
@@ -1,4 +1,13 @@
 <div id="<%= dom_id borrowing %>">
+
+  <p>
+    <strong>Game:</strong>
+    <%= borrowing.game.name %>
+  </p>
+  <p>
+    <strong>User:</strong>
+    <%= borrowing.user.fullname %>
+  </p>
   <p>
     <strong>Starts on:</strong>
     <%= borrowing.starts_on %>
