+++ app/views/admin/borrowings/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(model: borrowing) do |form| %>
+<%= form_with(model: [:admin, borrowing]) do |form| %>
   <% if borrowing.errors.any? %>
     <div style="color: red">
       <h2><%= pluralize(borrowing.errors.count, "error") %> prohibited this borrowing from being saved:</h2>
@@ -11,6 +11,14 @@
     </div>
   <% end %>
 
+  <div>
+    <%= form.label :game_id, style: "display: block" %>
+    <%= form.select :game_id, Game.all.map{|s| [s.name,s.id]} %>
+  </div>
+  <div>
+    <%= form.label :user_id, style: "display: block" %>
+    <%= form.select :user_id, User.all.map{|s| [s.fullname,s.id]} %>
+  </div>
   <div>
     <%= form.label :starts_on, style: "display: block" %>
     <%= form.date_field :starts_on %>
