+++ app/views/admin/borrowings/edit.html.erb
@@ -1,0 +1,10 @@
+<h1>Editing borrowing</h1>
+
+<%= render "form", borrowing: @borrowing %>
+
+<br>
+
+<div>
+  <%= link_to "Show this borrowing", admin_borrowing_url(@borrowing) %> |
+  <%= link_to "Back to borrowings", admin_borrowings_path %>
+</div>
