+++ app/views/admin/borrowings/new.html.erb
@@ -5,5 +5,5 @@
 <br>
 
 <div>
-  <%= link_to "Back to borrowings", borrowings_path %>
+  <%= link_to "Back to borrowings", admin_borrowings_path %>
 </div>
