+++ app/views/admin/borrowings/show.html.erb
@@ -1,0 +1,10 @@
+<p style="color: green"><%= notice %></p>
+
+<%= render @borrowing %>
+
+<div>
+  <%= link_to "Edit this borrowing", edit_admin_borrowing_path(@borrowing) %> |
+  <%= link_to "Back to borrowings", admin_borrowings_path %>
+
+  <%= button_to "Destroy this borrowing", @admin_borrowing, method: :delete %>
+</div>
