--- app/views/borrowings/edit.html.erb
@@ -1,10 +1,0 @@
-<h1>Editing borrowing</h1>
-
-<%= render "form", borrowing: @borrowing %>
-
-<br>
-
-<div>
-  <%= link_to "Show this borrowing", @borrowing %> |
-  <%= link_to "Back to borrowings", borrowings_path %>
-</div>
