+++ app/views/home/index.html.erb
@@ -2,5 +2,5 @@
 <% if user_signed_in? %>
   <%= link_to 'Users', admin_users_path %>
   <%= link_to 'Games', games_path %> 
-  <%= link_to 'Borrowings', borrowings_path %> 
+  <%= link_to 'Borrowings', admin_borrowings_path %> 
 <% end %>
